+++ src/DriveDrop.Web/DriveDrop.Web/ViewModels/Customer.cs
@@ -27,8 +27,8 @@ namespace DriveDrop.Web.ViewModels
         public CustomerStatus CustomerStatus { get;  set; }
         //public string CustomerStatus { get;  set; }
         public int? MaxPackage { get;  set; }
-        public int? PickupRadius { get;  set; }
-        public int? DeliverRadius { get;  set; }
+        public double? PickupRadius { get;  set; }
+        public double? DeliverRadius { get;  set; }
         public string PersonalPhotoUri { get;  set; }
         public string VehicleInfo { get; set; }
